requirements.txt
streamlit
pyautogen
langchain
langchain_community
requests
beautifulsoup4
markdownify
youtube-transcript-api
tavily-python